+++ src/contentscripts/actions/next-episode.js
@@ -1,5 +1,4 @@
-let nextEpisodeObserver;
-let currentEpisode;
+let legacyNextEpisodeObserver, nextEpisodeObserver, currentEpisode;
 
 /**
  * Listen to any action that triggers a pause (or triggered by other party members)
@@ -18,21 +17,38 @@ listenToWindowEvent('next-episode', async (ev) => {
  */
 function startNextEpisodeListener() {
     if (!player) return false;
+
+    currentEpisode = getSeasonAndEpisode();
+
+    // The legacy web player completely replaces the video element with a different one.
+    legacyNextEpisodeObserver?.disconnect();
+    legacyNextEpisodeObserver = new MutationObserver(handlePlayerParentChange);
+    const legacyObserveConfig = {childList: true, characterData: false};
+    legacyNextEpisodeObserver.observe(jQuery(player).parent()[0], legacyObserveConfig);
+
+    // The new web player simply changes the video `src` attribute
     nextEpisodeObserver?.disconnect();
     nextEpisodeObserver = new MutationObserver(handlePlayerParentChange);
-    const observeConfig =  {
-        attributes: false,
-        childList: true,
-        characterData: false
-    };
-    currentEpisode = getSeasonAndEpisode();
-    nextEpisodeObserver.observe(jQuery(player).parent()[0], observeConfig);
+    const observeConfig = {attributes: true, characterData: false};
+    nextEpisodeObserver.observe(jQuery(player)[0], observeConfig);
+}
+
+function stopNextEpisodeListener() {
+    legacyNextEpisodeObserver?.disconnect();
+    nextEpisodeObserver?.disconnect();
 }
 
 function handlePlayerParentChange() {
-    if (webPlayerIsOpen() && setPlayer() === undefined && currentEpisode) {
-        nextEpisodeObserver.disconnect();
-        // Perform a hacky trick by adding 'remote', so we do not emit it through socket.
-        postWindowMessage({type: 'start-video', reason: 'next-episode', remote: true});
-    }
+    // Use a timeout to give the closeObserver some time to check if the webplayer has closed.
+    setTimeout(() => {
+        if (webPlayerIsOpen() && player && currentEpisode) {
+            console.log('Next episode detected!');
+            player = getPlayer();
+            if (player === undefined) {
+                stopNextEpisodeListener();
+            }
+            // Perform a hacky trick by adding 'remote', so we do not emit it through socket.
+            postWindowMessage({type: 'start-video', reason: 'next-episode', remote: true});
+        }
+    }, 200);
 }