+++ src/contentscripts/actions/seek.js
@@ -4,15 +4,14 @@ let postponedSeekToTime;
  * Listen to any event that triggers a jump to another time (can also be triggered by other party members)
  */
 listenToWindowEvent('seek-video', async (ev) => {
-    if (ev.data.isLegacyPlayer && !isLegacyWebPlayer() && ev.data.remote) ev.data.time -= 10;
-    else if (!ev.data.isLegacyPlayer && isLegacyWebPlayer() && ev.data.remote) ev.data.time += 10;
+    const time = ev.data.time += currentTimeOffset;
 
     if (isPlayingTrailer()) {
         // In case someone seeks when we are still watching
         // a trailer, postpone the seek until later.
-        postponedSeekToTime = ev.data.time;
+        postponedSeekToTime = time;
     } else {
-        performSeek(ev.data.time);
+        performSeek(time);
     }
 });
 
@@ -24,7 +23,7 @@ function performSeek(time, emitPlayerReady = true) {
     if (time === undefined) return;
     player.onseeked = () => {
         if (emitPlayerReady) {
-            postWindowMessage({type: 'player-ready'});
+            postWindowMessage({type: 'player-ready'}); // Server will also emit a state update
         }
         player.onseeked = onSeeked;
     };
@@ -37,5 +36,5 @@ function performSeek(time, emitPlayerReady = true) {
  */
 function onSeeked() {
     if (isPlayingTrailer() || signalReadiness || !player) return;
-    postWindowMessage({type: 'seek-video', time: player.currentTime});
+    postWindowMessage({type: 'seek-video', time: player.currentTime - currentTimeOffset});
 }