+++ src/contentscripts/listener/websocket.js
@@ -4,7 +4,7 @@
 
 const supportedUi = '3.19.4-2020-05-14'; // Amazon user interface version. Different versions might imply webplayer changes.
 const websocketUrl = 'https://ws.primevideoparty.com';
-let currentParty; // = {id: string, members: string[], videoId: string}
+let currentParty; // = {id: string, members: {id: string, displayName: string}[], videoId: string}
 let socket;
 let displayName;
 
@@ -19,6 +19,11 @@ window.addEventListener('message', async function (ev) {
     if (ev.data.type === 'party-info' && ev.data.isNew) {
         // In case the user opens the extension for the first time in browser session,
         // or when the user clicks the 'new party' button.
+
+        if (currentParty && socket) {
+            socket.emit('leave-party', {});
+        }
+
         let baseUrl = '';
         if (window.isOnAmazonWebsite) {
             baseUrl = '/gp/video/storefront';
@@ -69,8 +74,12 @@ window.addEventListener('message', async function (ev) {
         case 'pause-video':
             socket.emit('pause-video', {
                 time: ev.data.time,
-                isLegacyPlayer: isLegacyWebPlayer()
+                isLegacyPlayer: isLegacyWebPlayer(),
+                reason: ev.data.reason
             });
+            break;
+        case 'state-update':
+            socket.emit('state-update', {state: ev.data.state});
             break;
         case 'play-video':
             socket.emit('play-video');
@@ -110,17 +119,32 @@ function initializeWebsocket(partyId) {
         postWindowMessage({
             type: 'play-video',
             coordinated: data.coordinated,
-            byMemberName: data.byMemberName,
+            byMember: data.byMember,
+            remote: true
+        })
+    });
+    socket.on('state-update', (data) => {
+        postWindowMessage({
+            type: 'state-update',
+            byMember: data.byMember,
+            state: data.state,
             remote: true
         })
     });
     socket.on('join-party', (data) => {
         currentParty.members = data.currentMembers;
+        for (let member of currentParty.members) {
+            if (member.id !== socket.id) continue;
+            member.isMe = true;
+            break;
+        }
+
         postWindowMessage({
             type: 'member-change',
             change: 'join',
             member: data.member,
             pause: data.pause,
+            currentParty,
             remote: true
         })
     });
@@ -130,13 +154,14 @@ function initializeWebsocket(partyId) {
             type: 'member-change',
             change: 'leave',
             member: data.member,
+            currentParty,
             remote: true
         })
     });
     socket.on('pause-video', (data) => {
         postWindowMessage({
             type: 'pause-video',
-            byMemberName: data.byMemberName,
+            byMember: data.byMember,
             time: data.time,
             isLegacyPlayer: data.isLegacyPlayer,
             remote: true
@@ -145,36 +170,40 @@ function initializeWebsocket(partyId) {
     socket.on('next-episode', (data) => {
         postWindowMessage({
             type: 'next-episode',
-            byMemberName: data.byMemberName,
+            byMember: data.byMember,
             season: data.season,
             episode: data.episode,
             remote: true
-        })
+        });
     });
     socket.on('watching-trailer', (data) => {
         postWindowMessage({
             type: 'watching-trailer',
-            byMemberName: data.byMemberName,
+            byMember: data.byMember,
             remote: true
         })
     });
     socket.on('seek-video', (data) => {
         postWindowMessage({
             type: 'seek-video',
-            byMemberName: data.byMemberName,
+            byMember: data.byMember,
             time: data.time,
             isLegacyPlayer: data.isLegacyPlayer,
             remote: true
         })
     });
     socket.on('close-video', (data) => {
-        postWindowMessage({type: 'close-video', byMemberName: data.byMemberName, remote: true})
+        postWindowMessage({
+            type: 'close-video',
+            byMember: data.byMember,
+            remote: true
+        })
     });
     socket.on('start-video-for-member', (data) => {
         // The server is asking us for the current time so another member can join in sync
         // If the currentTime in legacy webplayer = -10 seconds for new webplayer
         data.isLegacyPlayer = isLegacyWebPlayer();
-        data.time = player ? player.currentTime : 0;
+        data.time = player ? player.currentTime - currentTimeOffset : 0;
         socket.emit('start-video-for-member', data);
     });
     socket.on('start-video', (data) => {
@@ -189,7 +218,7 @@ function initializeWebsocket(partyId) {
                 ref: data.ref,
                 time: data.time,
                 isLegacyPlayer: data.isLegacyPlayer,
-                byMemberName: data.byMemberName,
+                byMember: data.byMember,
                 remote: true
             });
         } else {
@@ -199,6 +228,13 @@ function initializeWebsocket(partyId) {
                 urlBase = new URL(window.location).origin;
             }
             window.location.href = urlBase + '/detail/' + data.videoId + '/ref=' + data.ref + '?autoplay=1&t=' + startTime;
+        }
+    });
+
+    // Socket error handling
+    socket.on('reconnecting', (attemptNr) => {
+        if (attemptNr === 1) {
+            sendNotification('error', 'The server might be in the process of updating. Sorry for the inconvenience.', 'Lost connection');
         }
     });
 }